requirements.txt
asyncio
python-dotenv
telethon
